most popular product

+++ internal/models/product.go
@@ -5,5 +5,5 @@ import "github.com/google/uuid"
 type Product struct {
 	ID       uuid.UUID `json:"id"`
 	Name     string    `json:"name"`
-	Quantity int32     `json:"quantity"`
+	Quantity int64     `json:"quantity"`
 }
